spec: find system python via meson
RH-Author: Paolo Bonzini <pbonzini@redhat.com>
Message-id: <20201216175308.1463822-2-pbonzini@redhat.com>
Patchwork-id: 100482
O-Subject: [RHEL-AV-8.4.0 qemu-kvm PATCH 1/2] spec: find system python via meson
Bugzilla: 1899619
RH-Acked-by: Philippe Mathieu-Daudé <philmd@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Danilo de Paula <ddepaula@redhat.com>

When using --python, QEMU always uses the embedded copy of meson
instead of the system one.  It does this because --python
means "use this Python 3 for everything including Meson", and
on some systems it is not possible to pick a Python 3 to run the
system meson with.

In order to use the system meson and pick up any bugfixes
automatically, do not use --python.  Meson will pick the
system installation of Python from its shebang line and use it.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>