redhat: makes qemu respect system's crypto profile
RH-Author: Danilo de Paula <ddepaula@redhat.com>
Message-id: <20210204025545.3423037-2-ddepaula@redhat.com>
Patchwork-id: 100976
O-Subject: [RHEL-AV-8.4.0 RHEL-8.4.0 qemu-kvm PATCH 1/1] redhat: makes qemu respect system's crypto profile
Bugzilla: 1902960
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Philippe Mathieu-Daudé <philmd@redhat.com>
RH-Acked-by: Daniel P. Berrange <berrange@redhat.com>

When RHEL-8's (and AV's) qemu-kvm was created, it was based in RHEL-7's
qemu and not in Fedora.
There's no concept of cryto policy in RHEL-7 so qemu-kvm was using
--tls-priority=NORMAL back then.

In RHEL-8 and AV it should be using --tls-priority=@QEMU,SYSTEM as in
Fedora.

This would make QEMU use the "@QEMU" crypto policy profile, and if that
doesn't exist, then it would use the @SYSTEM crypto policy profile.
The upshot of using "NORMAL" is that QEMU is exposing the GNUTLS
defaults which are undesirable because they still allow TLS-1.0 and
TLS-1.1

The DEFAULT RHEL8 crypto policies only allow TLS-1.2

Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>