Limit build on Power to qemu-img and qemu-ga only
RH-Author: Miroslav Rezanina (mrezanin)
RH-MergeRequest: 1: Limit build on Power to qemu-img and qemu-ga only
RH-Commit: [1/1] 382a2e0b046ad8772b14919c8f51804851b707ba (mrezanin/rhel-src-qemu-kvm)
RH-Bugzilla: 1944056
RH-Acked-by: Danilo C L de Paula <ddepaula@redhat.com>
RH-Acked-by: David Gibson <dgibson@redhat.com>

As we do not support power as host on RHEL 9 we want to disable build of
qemu-kvm to limit supported code.

However, we need to support RHEL 9 as guest so we need qemu-ga and qemu-img
packages to be available. This is the reason we can't just disable power
build.

Updating specfile so we build only required subpackages.

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>