Remove message with running VM count
RH-Author: Miroslav Rezanina (mrezanin)
RH-MergeRequest: 2: Remove message with running VM count
RH-Commit: [1/1] 920a625530ef99f7d40fc0d0f2e90651f5a694f6 (mrezanin/rhel-src-qemu-kvm)
RH-Bugzilla: 1914461
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Danilo Cesar <ddepaula@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

In RHEL 8.4.0 we removed warning for exceeding VM count as limit
was removed in RHEL. However, we kept notification when count of
running VMs change.

We want to completely remove this message in RHEL 9. So we remove
check script and udev rule running it.

Signed-off-by: Miroslav Rezanina <mezanin@redhat.com>