redhat: Install the s390-netboot.img that we've built
RH-Author: Thomas Huth <thuth@redhat.com>
Message-id: <20210601134410.271683-2-thuth@redhat.com>
Patchwork-id: 101659
O-Subject: [RHEL-AV-8.5.0 qemu-kvm PATCH 1/1] redhat: Install the s390-netboot.img that we've built
Bugzilla: 1966463
RH-Acked-by: David Hildenbrand <david@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>
RH-Acked-by: Philippe Mathieu-Daudé <philmd@redhat.com>

Upstream-status: n/a

So far we've always been providing the pre-built s390-netboot.img that
ships with the QEMU release which we were using as our base. This is
unfortunate, since we can not use downstream patches for this binary
this way. Let's rather use the s390-netboot.img that is generated during
the build process.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>