redhat: Install the s390-netboot.img that we've built
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 8: Synchronize with RHEL-AV 8.5 release 19 to RHEL 9
RH-Commit: [2/8] 5d7ab7900cd6157697b8a5c8f58ff6e8951285af (mrezanin/centos-src-qemu-kvm)
RH-Bugzilla: 1957194
RH-Acked-by: Daniel P. Berrangé <berrange@redhat.com>
RH-Acked-by: Greg Kurz <gkurz@redhat.com>
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>

Upstream-status: n/a

So far we've always been providing the pre-built s390-netboot.img that
ships with the QEMU release which we were using as our base. This is
unfortunate, since we can not use downstream patches for this binary
this way. Let's rather use the s390-netboot.img that is generated during
the build process.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>