qga: drop StandardError=syslog
RH-Author: Marc-André Lureau <None>
RH-MergeRequest: 11: qga: drop StandardError=syslog
RH-Commit: [1/1] 2d6242b83498040593bb48140109da13cf20232b
RH-Bugzilla: 1947977
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Daniel P. Berrangé <berrange@redhat.com>

Since systemd v246:

 * StandardError= and StandardOutput= in unit files no longer support
   the "syslog" and "syslog-console" switches. They were long removed from
   the documentation, but will now result in warnings when used, and be
   converted to "journal" and "journal+console" automatically.

Note for reviewers, when 'StandardError' setting is omitted, it inherits
the same FD as for 'StandardOutput' and the latter in turn defaults to
value of the global setting 'DefaultStandardError' which in turn
defaults to 'journal'. IOW, simply deleting this setting has no
functional change on where stderr ends up in an out of the box setup.

Upstream: N/A
Resolves: rhbz#1947977

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>