spec: %files: don't use fine grained -docs file list
RH-Author: Cole Robinson <crobinso@redhat.com>
RH-MergeRequest: 12: spec: more reorg to make it easier to share with Fedora
RH-Commit: [19/21] e9e151b707694f29064352fa4e12172855891999 (crobinso/qemu-kvm-srcgit)
RH-Bugzilla: 1973029
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

This simplifies sharing with Fedora, which doesn't have the
README.systemtap file. But also I think it's fine to do this for
docs anyways: new stuff that pops up there we will likely want to
ship by default, and maintaining a whitelist slows down the rebase
process.

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>