redhat: Move qemu-kvm-docs dependency to qemu-kvm
RH-Author: Danilo de Paula <ddepaula@redhat.com>
Message-id: <20210629170812.2397650-3-ddepaula@redhat.com>
Patchwork-id: 101797
O-Subject: [RHEL-AV-8.5.0 qemu-kvm PATCH 2/3] redhat: Move qemu-kvm-docs dependency to qemu-kvm
Bugzilla: 1950405 1967330
RH-Acked-by: Philippe Mathieu-Daudé <philmd@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>

qemu-kvm-docs weak dependency was introduced to qemu-kvm-common
in order to make it easier for users who installed qemu-kvm-common
directly to do not miss docs after the sub package split.

Since that, new dependencies were introduced (like the spice-ui
submodule) and a discussion went on and it was agreed that strong
dependency on qemu-kvm was better than a weak dependency on the
core package.

This commit makes qemu-kvm-docs behave as the other recently
introduced subpackages.

Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>