redhat: introducting qemu-kvm-hw-usbredir
RH-Author: Danilo de Paula <ddepaula@redhat.com>
Message-id: <20210629170812.2397650-4-ddepaula@redhat.com>
Patchwork-id: 101794
O-Subject: [RHEL-AV-8.5.0 qemu-kvm PATCH 3/3] redhat: introducting qemu-kvm-hw-usbredir
Bugzilla: 1950405 1967330
RH-Acked-by: Philippe Mathieu-Daudé <philmd@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>

This submodule allows the removal of the usbredir dependency
from qemu-kvm-core, allowing it to be installed without usbredir
support.

This is part of the work to make qemu-kvm-core run with
the lightest footprint possible. That work started with the split
of the spice-ui and docs subpackages.

Worth noticing that the qemu-kvm package still depends on
qemu-kvm-hw-usbredir.

Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>