redhat: Replace the kvm-setup.service with a /etc/modules-load.d config file
RH-Author: Thomas Huth <thuth@redhat.com>
RH-MergeRequest: 18: Replace the kvm-setup.service with a /etc/modules-load.d config file
RH-Commit: [1/1] 0885ea45a87c7c6a5578b9832617dbe339c13f12 (thuth/qemu-kvm-cs9)
RH-Bugzilla: 1978837
RH-Acked-by: Daniel P. Berrangé <berrange@redhat.com>
RH-Acked-by: Paolo Bonzini <None>
RH-Acked-by: David Hildenbrand <david@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>

POWER won't be supported in RHEL9, so most parts of the kvm-setup script
are not required anymore. We still need a way to automatically load the
kvm.ko kernel module after a reboot on s390x, but that can be done more
easily with a config file in /etc/modules-load.d.

This patch now also removes the old "kvm.modules" - it's not installed
by the spec file since quite a while anymore, thus it's unused, but
apparently it has been forgotten to be removed in the past.

And while we're at it, also remove the udev event magic from the spec
file. This was needed in the past to get the permissions of the /dev/kvm
file right, which however is not needed here anymore since this logic
has been moved to systemd and the kernel module instead (browse to
https://github.com/systemd/systemd/pull/7795 for more information).

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>