Split qemu-pr-helper to separate package
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 20: Split unnecessary utilities to separate packages
RH-Commit: [2/2] 3fef0d1460bdb8e79d8622f1d4379b448980945e (mrezanin/centos-src-qemu-kvm)
RH-Bugzilla: 1972300
RH-Acked-by: Danilo Cesar Lemes de Paula <ddepaula@redhat.com>
RH-Acked-by: Philippe Mathieu-Daudé <philmd@redhat.com>
RH-Acked-by: Daniel P. Berrangé <berrange@redhat.com>
RH-Acked-by: Cole Robinson <crobinso@redhat.com>

qemu-pr-helper is required for certian SCSI operations. We do not need it
in core installation so we are moving it to separate package to lower
dependecies for core installation.

qemu-kvm-tools package is not use for qemu-pr-helper as it can be
required for qemu-kvm runtime (unline other utilities in qemu-kvm-tools).

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>