redhat: Move qemu-kvm-docs dependency to qemu-kvm
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 22: Synchronize with RHEL-AV 8.5 release 23 to RHEL 9
RH-Commit: [3/5] 82e92a0ab881e063330372d0ce78d6d4c0012a97 (mrezanin/centos-src-qemu-kvm)
RH-Bugzilla: 1957194

qemu-kvm-docs weak dependency was introduced to qemu-kvm-common
in order to make it easier for users who installed qemu-kvm-common
directly to do not miss docs after the sub package split.

Since that, new dependencies were introduced (like the spice-ui
submodule) and a discussion went on and it was agreed that strong
dependency on qemu-kvm was better than a weak dependency on the
core package.

This commit makes qemu-kvm-docs behave as the other recently
introduced subpackages.

Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>