redhat: introducting qemu-kvm-hw-usbredir
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 22: Synchronize with RHEL-AV 8.5 release 23 to RHEL 9
RH-Commit: [4/5] 748d4f0b3b2c3feeecd8f4240723ba60626d7723 (mrezanin/centos-src-qemu-kvm)
RH-Bugzilla: 1957194

This submodule allows the removal of the usbredir dependency
from qemu-kvm-core, allowing it to be installed without usbredir
support.

This is part of the work to make qemu-kvm-core run with
the lightest footprint possible. That work started with the split
of the spice-ui and docs subpackages.

Worth noticing that the qemu-kvm package still depends on
qemu-kvm-hw-usbredir.

Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>