spec: Use Fedora's pattern for specifying rc version
RH-Author: Cole Robinson <crobinso@redhat.com>
RH-MergeRequest: 12: spec: more reorg to make it easier to share with Fedora
RH-Commit: [18/21] 164780d21e94a83b4313a0c65cfc67ec6e9093e4 (crobinso/qemu-kvm-srcgit)
RH-Bugzilla: 1973029
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

Set a '%global rcver rc4' at the top of the file, and %setup
and URL fields are filled in as appropriate.

This also adds the -q option to %setup, to match Fedora. Most of
RHEL's history used this option, but it was dropped in RHEL8 timeframe.
I'm not sure if it was intentional, but in my experience it makes
build logs way way longer with out adding much value

Signed-off-by: Cole Robinson <crobinso@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>